tiled_movement_system deleted and it's content moved to movement_system. No need for any abstraction because only the tiled movement system will be used. #138
