+++ src/game/snake/movement_system.h
@@ -3,7 +3,10 @@
 
 #include <vector>
 #include <engine/component/component.h>
+#include <engine/component/initializable.h>
+#include <engine/entity.h>
 #include <engine/vector2.h>
+#include <game/tile_maze/tile_maze.h>
 
 struct segment_correction
 {
@@ -17,12 +20,17 @@ struct turning_correction
     vector2 head_direction;
 };
 
-class movement_system : public component
+class movement_system : public component, public initializable
 {
 public:
     movement_system(entity &attached_to);
-    virtual segment_correction correct_segments(const vector2 &start, const vector2 &end) const = 0;
-    virtual turning_correction correct_turning(const vector2& head_position, const vector2 &turn_direction) const = 0;
+    void initialize() override;
+    segment_correction correct_segments(const vector2 &start, const vector2 &end) const;
+    turning_correction correct_turning(const vector2& head_position, const vector2 &turn_direction) const;
+private:
+    vector2 project_longer_axis(const vector2 &position) const;
+
+    const tile_maze *_tile_maze;
 };
 
 #endif
